Move tllog orga views
