+++ bp/tllogs/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path, include
 
 from bp.views import APILogMarkReadView, APILogMarkHandledView, APILogMarkGoodView, APILogMarkBadView
-from bp.views import LogListView, LogAttentionListView, LogUnreadListView, LogReminderView, LogView
+from .orga.views import LogListView, LogAttentionListView, LogUnreadListView, LogReminderView, LogView
 from .tl.views import LogTLOverview, LogTLCreateView, LogTLUpdateView, LogTLDeleteView, LogTLDetailView
 
 tllog_patterns = [
